requirements.txt
pybaseball>=2.2.5
yahoo-fantasy-api>=1.9.0
requests>=2.31.0
jinja2>=3.1.2
sendgrid>=6.10.0
python-dotenv>=1.0.0
